tests/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
attrs==24.2.0
bidict==0.23.1
blinker==1.9.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.8
contourpy==1.2.1
cycler==0.12.1
facenet-pytorch==2.6.0
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SocketIO==5.5.1
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.12.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.65.4
h11==0.14.0
h5py==3.11.0
idna==3.7
importlib_metadata==8.2.0
importlib_resources==6.4.0
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.5
joblib==1.4.2
keras==3.4.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
mdurl==0.1.2
mediapipe==0.10.14
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
networkx==3.2.1
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.3
pillow==10.2.0
protobuf==4.25.4
pybind11==2.13.1
pycparser==2.22
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2024.2
requests==2.32.3
rich==13.7.1
scikit-learn==1.6.0
scikit-plot==0.3.7
scipy==1.11.4
seaborn==0.13.2
simple-websocket==1.1.0
six==1.16.0
sounddevice==0.4.7
sympy==1.13.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
tensorflow-metal==1.1.0
termcolor==2.4.0
threadpoolctl==3.5.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
Werkzeug==3.1.3
wrapt==1.16.0
wsproto==1.2.0
zipp==3.19.2
